Adapt the extract img file name with multiple same file types
There are multiple same file types with name 'SUPER' appears in the latest 'UPDATE.APP' file,
which is from the HM4 OS. There may be multiple other same file types in the future.
So add the extract img file name with the index of the file block can avoid the issue.

Signed-off-by: zoujunhua <zoujunhua86@gmail.com>